change group id data type from int to decimal. add new 3 columns in application user.

+++ DPEMoveDAL/Models/MAccountType.cs
@@ -7,7 +7,7 @@ namespace DPEMoveDAL.Models
     {
         public string AccountTypeId { get; set; }
         public string AccountTypeName { get; set; }
-        public int? DefaultGroupId { get; set; }
+        public decimal? DefaultGroupId { get; set; }
         public string RequireProfileBoo { get; set; }
     }
 }
